feat: Games page on app router
