--- pages/games/styles.module.scss
@@ -1,96 +1,0 @@
-@import "../../styles/theme.scss";
-
-.background {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  background: $Secondary;
-  padding: 0 1em;
-}
-.games {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  width: 100%;
-}
-.list {
-  margin: 0;
-  padding: 0rem;
-  list-style: none;
-  width: 75%;
-  display: flex;
-  flex-direction: column;
-  li:nth-of-type(2n + 1) {
-    width: 100%;
-    display: flex;
-    justify-content: flex-start;
-  }
-  li:nth-of-type(2n) {
-    width: 100%;
-    display: flex;
-    justify-content: flex-end;
-  }
-}
-.title {
-  color: $Primary;
-  text-align: center;
-  font-weight: 700;
-  font-size: 2rem;
-  margin: 1rem 0;
-}
-.description {
-  text-align: center;
-  margin: 0;
-  color: $Paragraph;
-}
-
-.button {
-  border: none;
-  cursor: pointer;
-  background: $Accent;
-  font-size: 20px;
-  text-transform: capitalize;
-  padding: 0.5rem 1rem;
-  margin: 1rem;
-  border-radius: 30px;
-  color: $Secondary;
-  border: 1px solid $Accent;
-  &:hover {
-    background: $Secondary;
-    color: $Accent;
-  }
-}
-@media (max-width: 720px) {
-  .list {
-    width: 100%;
-    li:nth-of-type(2n + 1) {
-      width: 100%;
-      display: flex;
-      justify-content: center;
-    }
-    li:nth-of-type(2n) {
-      width: 100%;
-      display: flex;
-      justify-content: center;
-    }
-  }
-}
-
-@media (min-width: 1150px) {
-  .list {
-    display: grid;
-    grid-template-columns: 50% 50%;
-    width: 100%;
-    li:nth-of-type(2n + 1),
-    li:nth-of-type(2n) {
-      width: 100%;
-      display: flex;
-      justify-content: center;
-      grid-row: span 2;
-    }
-    li:nth-child(2) {
-      grid-row: 2 / span 2;
-      grid-column: 2;
-    }
-  }
-}
